pyproject.toml
[tool.poetry]
name = "investigation-phase"
version = "0.1.0"
description = ""
authors = ["Shunsuke Hidaka <gsus4onc@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
einops = "^0.3.0"
librosa = "^0.8.1"
matplotlib = "^3.4.2"
numpy = "^1.20.3"
pandas = "^1.2.4"
pyarrow = "^4.0.1"
pytorch-lightning = "^1.3.4"
seaborn = "^0.11.1"
SoundFile = "^0.10.3"
PyYAML = "^5.4.1"
torch = {url = "https://download.pytorch.org/whl/cu101/torch-1.8.1%2Bcu101-cp38-cp38-linux_x86_64.whl"}

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
